fix: 🐛 category file and fix bug
