+++ assets/js/cartPage/index.js
@@ -1,0 +1,4 @@
+import "./cart.js";
+import "./coupon.js";
+import "./payment.js";
+import "./render.js";
