+++ pages/cart.html
@@ -557,7 +557,6 @@
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 
-    <script type="module" src="../assets/js/cart.js"></script>
-    <script type="module" src="../assets/js/cart_coupon.js"></script>
+    <script type="module" src="../assets/js/cartPage"></script>
   </body>
 </html>
